rename p ptr, and remove the with... options for the iterator
