+++ ptr/ptr.go
@@ -1,4 +1,4 @@
-package p
+package ptr
 
 // String returns a pointer to the string value passed in.
 func String(v string) *string {
